Update life.py
Massive Change:
+ No creatures or food are spawned at the start.
+ 500 Particles are spawned (these are currently visible but can be hidden by commenting out their draw function).

Particles are of 5 types, and combinations of particles hitting each other create the creatures. When creatures die, they dispose back into particles. However, this is only a midway update, I am intending to allow particles to bond into patterns of several before becoming creatures. This makes more sense, as a creature like big-red, which starts with a size of 10, should not be creatued with only 2 particles :p
Nor should it dispose into 2 particles.